Moved the org.mrgeo.mapreduce.job package to org.mrgeo.job
 - for #254
 - since we don't use map/reduce anymore
