+++ mrgeo-core/src/main/java/org/mrgeo/mapalgebra/MapAlgebraJob.java
@@ -23,8 +23,8 @@ import org.mrgeo.data.DataProviderFactory;
 import org.mrgeo.data.ProtectionLevelUtils;
 import org.mrgeo.data.ProviderProperties;
 import org.mrgeo.data.image.MrsImageDataProvider;
-import org.mrgeo.mapreduce.job.JobResults;
-import org.mrgeo.mapreduce.job.RunnableJob;
+import org.mrgeo.job.JobResults;
+import org.mrgeo.job.RunnableJob;
 import org.mrgeo.utils.HadoopUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
