+++ mrgeo-core/src/test/java/org/mrgeo/test/MapOpTestUtils.java
@@ -29,8 +29,8 @@ import org.mrgeo.image.MrsPyramid;
 import org.mrgeo.image.RasterTileMerger;
 import org.mrgeo.mapalgebra.MapAlgebra;
 import org.mrgeo.mapalgebra.parser.ParserException;
-import org.mrgeo.mapreduce.job.JobCancelledException;
-import org.mrgeo.mapreduce.job.JobFailedException;
+import org.mrgeo.job.JobCancelledException;
+import org.mrgeo.job.JobFailedException;
 import org.mrgeo.image.MrsPyramidMetadata;
 import org.mrgeo.utils.GDALJavaUtils;
 import org.mrgeo.utils.GDALUtils;
