+++ mrgeo-core/src/test/java/org/mrgeo/test/TestUtils.java
@@ -28,8 +28,8 @@ import org.mrgeo.hdfs.utils.HadoopFileUtils;
 import org.mrgeo.image.MrsImage;
 import org.mrgeo.image.RasterTileMerger;
 import org.mrgeo.mapalgebra.parser.ParserException;
-import org.mrgeo.mapreduce.job.JobCancelledException;
-import org.mrgeo.mapreduce.job.JobFailedException;
+import org.mrgeo.job.JobCancelledException;
+import org.mrgeo.job.JobFailedException;
 import org.mrgeo.image.MrsPyramidMetadata;
 import org.mrgeo.utils.*;
 import org.mrgeo.utils.tms.Bounds;
