+++ mrgeo-services/mrgeo-services-core/src/test/java/org/mrgeo/resources/mrspyramid/RasterResourceTest.java
@@ -40,8 +40,8 @@ import org.mrgeo.data.ProviderProperties;
 import org.mrgeo.image.MrsPyramid;
 import org.mrgeo.junit.UnitTest;
 import org.mrgeo.mapalgebra.MapAlgebraJob;
-import org.mrgeo.mapreduce.job.JobDetails;
-import org.mrgeo.mapreduce.job.JobManager;
+import org.mrgeo.job.JobDetails;
+import org.mrgeo.job.JobManager;
 import org.mrgeo.resources.job.JobInfoResponse;
 import org.mrgeo.services.mrspyramid.MrsPyramidService;
 import org.mrgeo.services.mrspyramid.rendering.ImageRenderer;
